+++ lua/obsidian/commands/sync.lua
@@ -15,6 +15,7 @@ return function(data)
     local choice = api.confirm "No vaults are configured for syncing. Do you want to run the setup wizard?"
     if choice == "Yes" then
       sync.setup()
+      return
     else
       return
     end